included the parsers in their corresponding modules
