+++ mock_processing/core/config.py
@@ -271,7 +271,7 @@ class Parser(object):
         return dump
 
 
-class Table_Parser(Parser):
+class TableParser(Parser):
 
     default = ParameterCollection(
         Parameter(
@@ -394,246 +394,3 @@ class Table_Parser(Parser):
         self._column_map = {}
         self._traverse_dict(self._config)
         return self._column_map
-
-
-class Photometry_Parser(Parser):
-
-    default = ParameterCollection(
-        Parameter(
-            "legacy", bool, False,
-            "use legacy mode (van den Busch et al. 2020)"),
-        Parameter(
-            "limit_sigma", float, 1.0,
-            "sigma value of the detection (magnitude) limit with respect to "
-            "the sky background"),
-        Parameter(
-            "no_detect_value", float, 99.0,
-            "magnitude value assigned to undetected galaxies"),
-        Parameter(
-            "SN_detect", float, 1.0,
-            "signal-to-noise ratio detection limit"),
-        Parameter(
-            "SN_floor", float, 0.2,
-            "numerical lower limit for signal-to-noise ratio"),
-        ParameterGroup(
-            "intrinsic",
-            Parameter(
-                "r_effective", str, "shape/R_effective",
-                "path of the effective radius column in the data store"),
-            Parameter(
-                "flux_frac", float, 0.5,
-                "this defines the effective radius by setting the fraction of "
-                "the total flux/luminosity for which the radius of the source "
-                "is computed"),
-            header=None),
-        ParameterListing(
-            "limits", float,
-            header=(
-                "numerical values for the magnitude limits, the keys must be "
-                "the same as in column map file used for "
-                "mocks_init_pipeline")),
-        ParameterListing(
-            "PSF", float,
-            header=(
-                "numerical values for the PSF FWHM in arcsec, the keys must "
-                "be the same as in column map file used for "
-                "mocks_init_pipeline")),
-        ParameterGroup(
-            "GAaP",
-            Parameter(
-                "aper_min", float, 0.7,
-                "GAaP lower aperture size limit in arcsec"),
-            Parameter(
-                "aper_max", float, 2.0,
-                "GAaP upper aperture size limit in arcsec"),
-            header=None),
-        ParameterGroup(
-            "SExtractor",
-            Parameter(
-                "phot_autoparams", float, 2.5,
-                "MAG_AUTO-like scaling factor for Petrosian radius, here "
-                "applied to intrinsic galaxy size derived from effective "
-                "radius"),
-            header=None),
-        header=(
-            "This configuration file is required for mocks_apertures and "
-            "mocks_photometry. It defines the free parameters of the aperture "
-            "and photometry methods as well as observational PSF sizes and "
-            "magnitude limits."))
-
-    def _run_checks(self):
-        if set(self.limits.keys()) != set(self.PSF.keys()):
-            message = "some filter(s) do not provide pairs of magnitude limit "
-            message += "and PSF size"
-            raise KeyError(message)
-
-    @property
-    def filter_names(self):
-        return tuple(sorted(self.limits.keys()))
-
-
-class Matching_Parser(Parser):
-
-    default = ParameterCollection(
-        Parameter(
-            "input", str, "...",
-            "path to input table, must be MemmapTable compatible",
-            parser=expand_path),
-        Parameter(
-            "normalize", bool, False,
-            "whether each feature space dimension is normalized by the "
-            "standard deviation of the values in the dimension"),
-        Parameter(
-            "max_dist", float, 1.0,
-            "maximum distance (Euclidean metric) to consider the nearest "
-            "neighbour a match, otherwise the fallback value is assigned"),
-        Parameter(
-            "every_n", int, 1,
-            "take a regular sample of the input data to speed up tree build "
-            "and query times"),
-        ParameterListing(
-            "features", str,
-            header=(
-                "Mapping between columns of the input data table and the "
-                "columns of the mock pipeline data store. The input features "
-                "are used to build the search tree, the data store features "
-                "are matched against tree. Key must be a valid column name in "
-                "the input table, value an existing column name in the data "
-                "store.")),
-        ParameterListing(
-            "observables", str,
-            header=(
-                "Mapping between columns of the input data table and the "
-                "columns of the mock pipeline data store. The observables of "
-                "the input data that are assigned to the nearest neighbour of "
-                "an object in the data store. Key must be a valid column name "
-                "in the input table, value a pair of output column name in "
-                "the data store and fall back value if distance > max_dist.")),
-        ParameterListing(
-            "fallback", str,
-            header=(
-                "Fallback value to assign if the nearest neighbour distance > "
-                "max_dist. If no value is provided the value of the "
-                "observable is assigned instead. Key must match one of the "
-                "keys in observables, value must have the same data type as "
-                "the observables.")),
-        ParameterListing(
-            "weights", str,
-            header=(
-                "Weights are used to scale the feature space along specific "
-                "axes. This is useful to give a higher or lower significance "
-                "to some features when matching. If no weight is provided for "
-                "a feature, no scaling is applied. Key must match one of the "
-                "keys in observables, value must have the same data type as "
-                "the observables.")),
-        header=(
-            "This configuration file is required for mocks_match_data. It "
-            "defines the features of the mock data and an external data set "
-            "to match. Based on the nearest neighbour in feature space, a set "
-            "of observables from the external data is assigned to the mock "
-            "data. Additional distance, cut-offs, fallback values and "
-            "weightes for each feature space dimension can be specified."))
-
-    def _run_checks(self):
-        # check that there are no fallback values with no matching observable
-        if set(self.observables.keys()) < set(self.fallback.keys()):
-            message = "fallback values are provided for undefined observables"
-            raise KeyError(message)
-        # check that there are no weights with no matching observable
-        if set(self.observables.keys()) < set(self.weights.keys()):
-            message = "weights are provided for undefined observables"
-            raise KeyError(message)
-
-    @property
-    def feature_names(self):
-        return tuple(sorted(self.features.keys()))
-
-    @property
-    def observable_names(self):
-        return tuple(sorted(self.observables.keys()))
-
-
-class BPZ_Parser(Parser):
-
-    default = ParameterCollection(
-        Parameter(
-            "BPZpath", str, "~/src/bpz-1.99.3",
-            "path of the bpz installation to use",
-            parser=expand_path),
-        Parameter(
-            "BPZenv", str, None,
-            "path to a python2 environment used when calling BPZ (leave blank "
-            "to use the default python2 interpreter)",
-            parser=expand_path),
-        Parameter(
-            "BPZtemp", str, None,
-            "temporary directory to use for bpz (I/O intense, e.g. a ram disk, "
-            "if blank, defaults to data store)",
-            parser=expand_path),
-        Parameter(
-            "flux", bool, False,
-            "whether the input columns are fluxes or magnitudes"),
-        Parameter(
-            "system", str, "AB",
-            "photometric system, must be \"AB\" or \"Vega\""),
-        ParameterListing(
-            "filters", str,
-            header=(
-                "Mapping between filter keys (same as used in the column map "
-                "file for mocks_init_pipeline) and paths to transmission "
-                "curve files compatible with BPZ (do not need to be in the "
-                "BPZpath/FILTER directory)."),
-            parser=expand_path),
-        ParameterGroup(
-            "prior",
-            Parameter(
-                "name", str, "hdfn_gen",
-                "template-magnitude prior module in the bpz_path directory "
-                "(filename: prior_[name].py)"),
-            Parameter(
-                "filter", str, "i",
-                "filter key (one of those listed in the [filters] section) "
-                "based on which the prior is evaluated"),
-            header=None),
-        ParameterGroup(
-            "templates",
-            Parameter(
-                "name", str, "CWWSB4",
-                "name of the template .list file in the bpz_path/SED "
-                "directory"),
-            Parameter(
-                "interpolation", int, 10,
-                "introduces n points of interpolation between the templates "
-                "in the color space"),
-            header=None),
-        ParameterGroup(
-            "likelihood",
-            Parameter(
-                "zmin", float, 0.01,
-                "minimum redshift probed"),
-            Parameter(
-                "zmax", float, 7.00,
-                "maximum redshift probed"),
-            Parameter(
-                "dz", float, 0.01,
-                "redshift resolution, intervals are logarithmic: (1+z)*dz"),
-            Parameter(
-                "odds", float, 0.68,
-                "redshift confidence limits"),
-            Parameter(
-                "min_rms", float, 0.0,
-                "intrinsic scatter of the photo-z in dz/(1+z)"),
-            header=None),
-        header=(
-            "This configuration file is required for mocks_BPZ. It defines "
-            "the transmission curves, galaxy templates and Bayesian prior "
-            "required to run the BPZ photometric reshift code."))
-
-    def _run_checks(self):
-        if self.prior["filter"] not in self.filters:
-            message = "prior filter is not included in the filter list: {:}"
-            raise KeyError(message.format(self.prior["filter"]))
-
-    @property
-    def filter_names(self):
-        return tuple(sorted(self.filters.keys()))
