+++ mock_processing/samples.py
@@ -6,6 +6,8 @@ import numpy as np
 from scipy.interpolate import interp1d, interp2d
 
 from .core.bitmask import BitMaskManager as BMM
+from .core.config import (LineComment, Parameter, ParameterGroup,
+                          ParameterCollection, Parser)
 from .core.parallel import Schedule
 from .core.utils import ProgressBar
 from .matching import DistributionEstimator
@@ -21,6 +23,7 @@ DENSITY_FILE_TEMPLATE = os.path.join(SUCCESS_RATE_DIR, "{:}.json")
 # register all known selections here
 PHOTOMETRIC_SELECTIONS = OrderedDict()
 SAMPLING_SELECTIONS = OrderedDict()
+SELECTION_PARSERS = OrderedDict()
 REGISTERED_SAMPLES = set()
 
 
@@ -34,10 +37,58 @@ def register(selector):
     elif issubclass(selector, Sampler):
         SAMPLING_SELECTIONS[selector.name] = selector
         REGISTERED_SAMPLES.update(PHOTOMETRIC_SELECTIONS.keys())
+    elif issubclass(selector, Parser):
+        SELECTION_PARSERS[selector.name] = selector
+        REGISTERED_SAMPLES.update(PHOTOMETRIC_SELECTIONS.keys())
     else:
-        message = "Object must be subclass of 'BaseSelection' or 'Sampler'"
+        message = "Object must be subclass of 'BaseSelection', 'Sampler' or "
+        message += "'Parser'"
         raise TypeError(message)
     return selector
+
+
+class DumpConfig(argparse.Action):
+
+    def __init__(self, *args, nargs=1,**kwargs):
+        super().__init__(*args, nargs=nargs, **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string, **kwargs):
+        sample = values[0]
+        try:
+            sample_parser = REGISTERED_SAMPLES[sample]
+        except KeyError:
+            message = "sample '{:}' does not implement a configuration"
+            raise parser.error(message.format(sample))
+        print(sample_parser.default)
+        parser.exit()
+
+
+def make_bitmask(sample):
+    param = Parameter(
+        "bitmask", str, "samples/" + sample,
+        "path at which the sample selection bit mask is saved in the data "
+        "store")
+    return param
+
+
+header = (
+    "This configuration file is required for mocks_select_sample. It "
+    "defines the output column name of the selection bit mask and the "
+    "data columns for the selection functions (such as magnitudes or "
+    "redshift).")
+selection_header = (
+    "Mapping form keyword argument names in selection function to "
+    "column path in the data store. Optional arguments can be "
+    "left blank.")
+param_redshift = Parameter("redshift", str, "...", "(observed) redshifts")
+param_sdss = {}
+for key in ("g", "r", "i", "Z", "Ks"):
+    param_sdss[key] = Parameter(
+        "mag_" + key, str, "...", "SDSS {:}-band magnitude".format(key))
+param_johnson = {}
+for key in ("B", "Rc", "Ic"):
+    param_johnson[key] = Parameter(
+        "mag_" + key, str, "...", "Johnson {:}-band magnitude".format(key))
 
 
 class Sampler(object):
@@ -186,6 +237,27 @@ class BaseSelection(object):
 
 
 @register
+class ParserKiDS(Parser):
+
+    name = "KiDS"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            Parameter(
+                "recal_weight", str, "lensing/recal_weight",
+                "path of lensing weight in the data store"),
+            LineComment(
+                "setting the following paramter will remove objects with "
+                "non-detections"),
+            Parameter(
+                "prior_magnitude", str, None,
+                "path of BPZ prior magnitude in the data store"),
+            header=selection_header),
+        header=header)
+
+
+@register
 class SelectKiDS(BaseSelection):
 
     name = "KiDS"
@@ -210,6 +282,23 @@ class SelectKiDS(BaseSelection):
 ###############################################################################
 #                            WIDE SURVEYS                                     #
 ###############################################################################
+
+
+@register
+class Parser2dFLenS(Parser):
+
+    name = "2dFLenS"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_sdss["g"],
+            param_sdss["r"],
+            param_sdss["i"],
+            param_sdss["Z"],
+            param_sdss["Ks"],
+            header=selection_header),
+        header=header)
 
 
 @register
@@ -277,6 +366,19 @@ class Select2dFLenS(BaseSelection):
 
 
 @register
+class ParserGAMA(Parser):
+
+    name = "GAMA"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_sdss["r"],
+            header=selection_header),
+        header=header)
+
+
+@register
 class SelectGAMA(BaseSelection):
 
     name = "GAMA"
@@ -292,6 +394,30 @@ class SelectGAMA(BaseSelection):
         BMM.set_bit(bitmask, self._bits[0], condition=is_selected)
         # update the master selection bit
         BMM.update_master(bitmask, sum(self._bits))
+
+
+@register
+class ParserSDSS(Parser):
+
+    name = "SDSS"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_sdss["g"],
+            param_sdss["r"],
+            param_sdss["i"],
+            Parameter(
+                "is_central", str, "environ/is_central",
+                "flag indicating if it is central host galaxy"),
+            Parameter(
+                "lmhalo", str, "environ/log_M_halo",
+                "logarithmic halo mass"),
+            Parameter(
+                "lmstellar", str, "environ/log_M_stellar",
+                "logarithmic stellar mass"),
+            header=selection_header),
+        header=header)
 
 
 @register
@@ -361,6 +487,23 @@ class SelectSDSS(BaseSelection):
 
 
 @register
+class ParserWiggleZ(Parser):
+
+    name = "WiggleZ"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_redshift,
+            param_sdss["g"],
+            param_sdss["r"],
+            param_sdss["i"],
+            param_sdss["Z"],
+            header=selection_header),
+        header=header)
+
+
+@register
 class SelectWiggleZ(BaseSelection):
 
     name = "WiggleZ"
@@ -412,6 +555,21 @@ class SampleWiggleZ(RedshiftSampler):
 ###############################################################################
 #                            DEEP SURVEYS                                     #
 ###############################################################################
+
+
+@register
+class ParserDEEP2(Parser):
+
+    name = "DEEP2"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_johnson["B"],
+            param_johnson["Rc"],
+            param_johnson["Ic"],
+            header=selection_header),
+        header=header)
 
 
 @register
@@ -479,6 +637,20 @@ class SampleDEEP2(DensitySampler):
     def __init__(self, bit_manager, mock_area, bitmask):
         density_file = DENSITY_FILE_TEMPLATE.format(self.name)
         super().__init__(bit_manager, density_file, mock_area, bitmask)
+
+
+@register
+class ParserVVDSf02(Parser):
+
+    name = "VVDSf02"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_redshift,
+            param_johnson["Ic"],
+            header=selection_header),
+        header=header)
 
 
 @register
@@ -562,6 +734,20 @@ class SampleVVDSf02(DensitySampler):
 
 
 @register
+class ParserzCOSMOS(Parser):
+
+    name = "zCOSMOS"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_redshift,
+            param_johnson["Ic"],
+            header=selection_header),
+        header=header)
+
+
+@register
 class SelectzCOSMOS(BaseSelection):
 
     name = "zCOSMOS"
@@ -624,6 +810,19 @@ class SamplezCOSMOS(DensitySampler):
 ###############################################################################
 #                            TEST SAMPLES                                     #
 ###############################################################################
+
+
+@register
+class ParserSparse24mag(Parser):
+
+    name = "Sparse24mag"
+    default = ParameterCollection(
+        make_bitmask(name),
+        ParameterGroup(
+            "selection",
+            param_sdss["r"],
+            header=selection_header),
+        header=header)
 
 
 @register
