+++ scripts/mocks_BPZ
@@ -4,15 +4,14 @@ import os
 import sys
 
 import mock_processing as mocks
-from mock_processing.core.config import BPZ_Parser
-from mock_processing.photoz import BpzManager
+from mock_processing.photoz import BpzManager, BpzParser
 
 
 parser = argparse.ArgumentParser(
     description="Compute magnitudes corrected for gravitational lensing "
                 "magnification in the limit of weak lensing.",
     add_help=False)
-parser.register("action", "dump", BPZ_Parser.get_dump())
+parser.register("action", "dump", BpzParser.get_dump())
 
 parser.add_argument(
     "datastore", help="directory in which the data store is located")
@@ -44,7 +43,7 @@ def main():
     logger = mocks.PipeLogger(__file__, args.datastore)
 
     # check the configuration file
-    config = BPZ_Parser(args.config, logger)
+    config = BpzParser(args.config, logger)
 
     # run BPZ on the selected magnitudes
     with mocks.DataStore.open(args.datastore, False, logger=logger) as ds:
