+++ scripts/mocks_photometry
@@ -4,15 +4,15 @@ import os
 import sys
 
 import mock_processing as mocks
-from mock_processing.core.config import Photometry_Parser
-from mock_processing.photometry import photometry_realisation_wrapped
+from mock_processing.photometry import (PhotometryParser,
+                                        photometry_realisation_wrapped)
 
 
 parser = argparse.ArgumentParser(
     description="Create a photometry realisation based on simulated apertures "
                 "and magnitude limits.",
     add_help=False)
-parser.register("action", "dump", Photometry_Parser.get_dump())
+parser.register("action", "dump", PhotometryParser.get_dump())
 
 parser.add_argument(
     "datastore", help="directory in which the data store is located")
@@ -59,7 +59,7 @@ def main():
     logger = mocks.PipeLogger(__file__, args.datastore)
 
     # check the configuration file
-    config = Photometry_Parser(args.config, logger)
+    config = PhotometryParser(args.config, logger)
 
     # apply the magnification correction to the model magnitudes
     with mocks.DataStore.open(args.datastore, False, logger=logger) as ds:
