+++ scripts/mocks_select_sample
@@ -6,14 +6,14 @@ import sys
 import mock_processing as mocks
 from mock_processing.core.bitmask import BitMaskManager
 from mock_processing.matching import DistributionEstimator
-from mock_processing import samples, sampleconfig
+from mock_processing import samples
 
 
 parser = argparse.ArgumentParser(
     description="Apply sample selections on to obtain a mock representation "
                 "of a specific photometric or spectroscopic sample.",
     add_help=False)
-parser.register("action", "dump", sampleconfig.DumpConfig)
+parser.register("action", "dump", samples.DumpConfig)
 
 parser.add_argument(
     "datastore", help="directory in which the data store is located")
@@ -43,7 +43,7 @@ help_group.add_argument(
     help="show this help message and exit")
 help_group.add_argument(
     "--dump", nargs=1, action="dump", metavar="sample",
-    choices=sorted(samples.REGISTERED_SAMPLES),
+    choices=sorted(samples.SELECTION_PARSERS),
     help="dump a default configuration file to stdout and exit for a given "
          "sample (see --sample for a list of valid names)")
 
@@ -54,7 +54,7 @@ def main():
     logger = mocks.PipeLogger(__file__, args.datastore)
 
     # check the configuration file
-    Parser = sampleconfig.REGISTERED_SAMPLES[args.sample]
+    Parser = samples.SELECTION_PARSERS[args.sample]
     config = Parser(args.config, logger)
 
     # apply the magnification correction to the model magnitudes
